refactor: move LocalEventBus to Providers subdirectory
Reorganize event bus architecture by moving LocalEventBus
class into dedicated Providers folder for better code
organization and separation of concerns.
