+++ MoLibrary.EventBus/Providers/LocalEventBus.cs
@@ -4,7 +4,7 @@ using MoLibrary.EventBus.Abstractions;
 using MoLibrary.EventBus.Abstractions.Handlers;
 using MoLibrary.EventBus.Abstractions.Subscriptions;
 
-namespace MoLibrary.EventBus;
+namespace MoLibrary.EventBus.Providers;
 
 /// <summary>
 /// Local (in-process) event bus implementation.
